requirements.txt
authlib>=1.0
accelerate==0.33.0
aiobotocore==2.19.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aioitertools==0.12.0
aiosignal==1.3.2
anndata==0.11.3
annotated-types==0.7.0
anyio==3.6.1
appdirs==1.4.4
appnope==0.1.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
array_api_compat==1.10.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.2
async-timeout==5.0.1
attrs==22.2.0
av==13.1.0
babel==2.12.1
beautifulsoup4==4.11.1
bleach==6.0.0
blinker==1.9.0
boto3==1.36.20
botocore==1.36.3
Brotli==1.0.9
cached-property==1.5.2
cachetools==5.3.0
cellxgene-census==1.15.0
certifi==2023.7.22
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.2.0
ChatterBot==1.2.0
click==8.1.8
cloup==3.0.5
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
Cython==3.0.12
dataclasses==0.6
datasets==2.19.2
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
et_xmlfile==2.0.1
evaluate==0.4.2
exceptiongroup==1.0.5
executing==1.2.0
fastjsonschema==2.16.2
ffmpeg==1.4
filelock==3.10.7
Flask==3.1.0
Flask-Cors==5.0.0
fonttools==4.39.4
fqdn==1.5.1
freetype-py==2.3.0
frozenlist==1.5.0
fsspec==2024.3.1
fuzzywuzzy==0.18.0
gget==0.29.0
glcontext==3.0.0
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
google_speech==1.2.0
googleapis-common-protos==1.67.0
greenlet==2.0.2
grpcio==1.70.0
grpcio-status==1.70.0
gTTS==2.5.4
h11==0.14.0
h2==4.0.0
h5py==3.12.1
hpack==4.0.0
httpcore==0.15.0
httplib2==0.22.0
httpx==0.24.1
huggingface-hub==0.28.1
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
ipykernel==6.22.0
ipython==8.12.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
json5==0.9.7
jsonpointer==2.0
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==5.3.0
jupyter-server==2.6.1
jupyterlab==4.0.0
jupyterlab-pygments==0.2.2
jupyterlab-server==2.20.0
jupyterlab-widgets==1.0.2
kiwisolver==1.4.4
legacy-api-wrap==1.4.1
llvmlite==0.44.0
lxml==5.3.1
manim==0.19.0
manim-voiceover==0.3.7
ManimPango==0.6.0
mapbox_earcut==1.0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.2
mathparse==0.1.2
matplotlib==3.10.0
matplotlib-inline==0.1.3
mdurl==0.1.2
mistune==2.0.5
moderngl==5.12.0
moderngl-window==3.1.1
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
munkres==1.1.4
mutagen==1.47.0
mysql-connector-python==9.2.0
natsort==8.4.0
nbclient==0.7.2
nbconvert==7.7.5
nbformat==5.9.3
nest-asyncio==1.5.6
networkx==3.4.2
notebook==6.5.3
notebook-shim==0.2.2
numba==0.61.0
numpy==1.23.5
openpyxl==3.1.5
overrides==6.2.0
packaging==23.1
pandas==2.2.2
pandocfilters==1.5.0
parso==0.8.3
patsy==1.0.1
pexpect==4.8.0
pickleshare==0.7.5
pillow==9.5.0
pkgutil-resolve-name==1.3.10
platformdirs==3.7.0
prometheus_client==0.16.0
prompt-toolkit==3.0.38
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
psutil==5.9.5
ptyprocess==0.7.0
pure_eval==0.2.2
pyarrow==12.0.1
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycairo==1.27.0
pycparser==2.21
pydantic==2.10.6
pydantic-core==2.27.2
pydub==0.25.1
pyglet==2.1.2
pyglm==2.8.0
Pygments==2.14.0
pynndescent==0.5.13
pyobjc-core==8.8.1
pyobjc-framework-Cocoa==8.8.1
pyparsing==3.1.2
pyserial==3.5
PySocks==1.7.1
pytdc==1.1.14
python-dateutil==2.8.2
python-dotenv==0.21.1
python-json-logger==2.0.2
python-slugify==8.0.4
python-vlc==3.0.21203
pytz==2022.7
PyYAML==6.0
pyzmq==25.1.1
rdkit==2023.9.6
referencing==1.0.0
regex==2024.11.6
reportlab==3.6.12
requests==2.28.1
rfc3339-validator==1.0.4
rfc3986-validator==0.2.1
rich==13.9.4
rlpycairo==0.1.5
rpds-py==0.7.2
rsa==4.9
s3fs==2024.3.1
s3transfer==0.11.2
safetensors==0.5.2
scanpy==1.11.0
scikit-learn==1.2.2
scipy==1.15.1
screeninfo==0.8.1
SDF==0.3.6
seaborn==0.13.2
Send2Trash==1.8.0
session-info2==0.1.2
six==1.16.0
skia-pathops==0.8.0.post2
sniffio==1.3.0
somacore==1.0.11
soupsieve==2.5
sox==1.5.0
SpeechRecognition==3.14.1
SQLAlchemy==2.0.30
srt==3.5.3
stack_data==0.6.0
statsmodels==0.14.4
svgelements==1.9.6
sympy==1.13.1
tdc==0.1
terminado==0.17.1
text-unidecode==1.3
threadpoolctl==3.4.0
tiledb==0.29.1
tiledbsoma==1.11.4
tinycss2==1.2.1
tokenizers==0.19.1
tomli==2.0.1
torch==2.6.0
tornado==6.3.3
tqdm==4.67.1
traitlets==5.9.0
transformers==4.43.4
types-python-dateutil==2.8.1
typing-extensions==4.10.0
typing_utils==0.4.0
tzdata==2022.7
umap-learn==0.5.7
unicodedata2==15.0.0
uri-template==4.1.1
uritemplate==4.1.1
urllib3==1.26.15
watchdog==6.0.0
wcwidth==0.2.6
web_cache==1.1.0
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==3.1.3
widgetsnbextension==3.6.0
wikipedia==1.4.0
wrapt==1.17.2
xgboost==1.7.4
xxhash==3.5.0
yapf==0.43.0
yarl==1.18.3
zipp==3.15.0
zstandard==0.21.0
>>>>>>> main
